backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
httpx
pydantic
python-dotenv
